move breadcrumbs to component
